pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "groq>=0.31.0",
    "langchain>=0.3.27",
    "langchain-groq>=0.3.7",
    "langserve>=0.3.1",
    "pinecone>=7.3.0",
    "python-dotenv>=1.1.1",
    "sse-starlette>=3.0.2",
    "streamlit>=1.48.1",
    "uvicorn>=0.35.0",
]
